# Conflicts:
#	src/SalesManager.java

+++ src/SalesManager.java
@@ -1,15 +1,14 @@
 public class SalesManager {
 
-    protected int[] sales;
+    protected long[] sales;
 
-    public SalesManager(int[] sales) {
-
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
@@ -17,9 +16,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
-        for (int sale : sales) {
+    public long min() {
+        long min = sales[0];
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -27,11 +26,11 @@ public class SalesManager {
         return min;
     }
 
-    public int truncatedMean() {
-        int sum = 0;
-        int kolvo = 0;
-        int max = max();
-        int min = min();
+    public long truncatedMean() {
+        long sum = 0;
+        long kolvo = 0;
+        long max = max();
+        long min = min();
 
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] != max && sales[i] != min) {
@@ -41,7 +40,7 @@ public class SalesManager {
         }
 
 
-        int result=0;
+        long result=0;
         if (kolvo != 0) {
             result = sum / kolvo;
         }
